Move downloading to goroutines Added download throttling Moved stats to separate file
